mapping coverage is a 32bits not as it arrived to be bigger than 256

+++ host/inc/genome.h
@@ -21,7 +21,7 @@
  */
 typedef struct {
     int8_t *data;
-    int8_t *mapping_coverage;
+    int32_t *mapping_coverage;
     int *substitution_list;
     uint32_t nb_seq;
     uint64_t pt_seq[MAX_SEQ_GEN];
